Phat overhaul of the system
Refactored the system to transfer events with generators and consumers. Each subsystem produces and consumes a certain set of events rather than all information being passed through the system in the form of a FloorEvent, like in iteration 1. The code currently does not function exactly properly and will need some troubleshooting. The scheduler class has also become quite bloated, and will need to be trimmed down later. Mainly just removing the event storage and retrieval methods and putting them in a separate class.