Pulled generateFloorEvent from the scheduler into its own thread, also added an ElevatorState class
Events are now generated in their own class, and only the required information is sent to the scheduler. There is also a class that keeps track of the state an elevator is in (current floor, destination, and direction) so it can be easily transferred to the scheduler for it to keep track of things.